+++ HISTORY.md
@@ -8,6 +8,7 @@
   code, they must be defined before use. The unfinished macro
   is good for that.
 * Added a leiningen plugin to get tidy test reports.
+* Chatty checkers
   
 = 0.6.1 (stable)
 * Midje-mode is (more) compatible with clojure-test-mode.